refactor api routes

+++ pages/api/categories/delete/[id].ts
@@ -1,10 +1,33 @@
 import type { NextApiRequest, NextApiResponse } from 'next';
-import { PrismaClient } from '@prisma/client';
+import { prisma } from '../../../../src/prisma';
 
-const prisma = new PrismaClient();
+export default async function handler(
+  req: NextApiRequest,
+  res: NextApiResponse
+) {
+  if (req.method !== 'POST') {
+    return res.status(405).end();
+  }
 
-async function deleteCategory(id: string | any) {
-  try {
+  const { id } = req.query;
+  if (typeof id === 'string') {
+    const categoryExists = await prisma.category
+      .findFirst({
+        where: {
+          id,
+        },
+      })
+      .finally(async () => {
+        await prisma.$disconnect();
+      });
+
+    if (!categoryExists) {
+      return res.status(404).json({
+        type: 'error',
+        response: 'Não existe nenhuma categoria com esse nome',
+      });
+    }
+
     const deleteAllArticlesOfCategory = prisma.articles.deleteMany({
       where: {
         categoryId: id,
@@ -15,30 +38,12 @@ async function deleteCategory(id: string | any) {
         id,
       },
     });
-    await prisma.$transaction([deleteAllArticlesOfCategory, deleteCategory]);
-    return;
-  } catch (error) {
-    // console.log(error)
-    return 'Ocorreu um erro ao deletar a categoria';
-  }
-}
-
-export default async function handler(
-  req: NextApiRequest,
-  res: NextApiResponse
-) {
-  if (req.method === 'POST') {
-    try {
-      const { id } = req.query;
-
-      await deleteCategory(id).finally(async () => {
+    await prisma
+      .$transaction([deleteAllArticlesOfCategory, deleteCategory])
+      .finally(async () => {
         await prisma.$disconnect();
       });
 
-      return res.status(200).json({ type: 'success' });
-    } catch (error) {
-      console.log(error);
-      return res.status(400).json({ type: 'error', response: error });
-    }
+    return res.status(200).json({ type: 'success' });
   }
 }
