+++ pages/api/clients/delete/[id].ts
@@ -1,10 +1,29 @@
 import type { NextApiRequest, NextApiResponse } from 'next';
-import { PrismaClient } from '@prisma/client';
+import { prisma } from '../../../../src/prisma';
 
-const prisma = new PrismaClient();
+export default async function handler(
+  req: NextApiRequest,
+  res: NextApiResponse
+) {
+  if (req.method === 'POST') {
+    return res.status(405).end();
+  }
+  const { id } = req.query;
 
-async function deleteClient(id: string | any) {
-  try {
+  if (typeof id === 'string') {
+    const clientExists = prisma.client.findFirst({
+      where: {
+        id,
+      },
+    });
+
+    if (!clientExists) {
+      return res.status(404).json({
+        type: 'error',
+        response: 'Não existe nenhum cliente com esse ID!',
+      });
+    }
+
     const deleteProjectsClient = prisma.project.deleteMany({
       where: {
         client_id: id,
@@ -17,30 +36,12 @@ async function deleteClient(id: string | any) {
       },
     });
 
-    await prisma.$transaction([deleteProjectsClient, deleteClient]);
-    return;
-  } catch (error) {
-    console.log(error);
-    return 'Ocorreu um erro ao deletar o cliente';
-  }
-}
-
-export default async function handler(
-  req: NextApiRequest,
-  res: NextApiResponse
-) {
-  if (req.method === 'POST') {
-    try {
-      const { id } = req.query;
-      console.log(id);
-      await deleteClient(id).finally(async () => {
+    await prisma
+      .$transaction([deleteProjectsClient, deleteClient])
+      .finally(async () => {
         await prisma.$disconnect();
       });
 
-      return res.status(200).json({ type: 'success' });
-    } catch (error) {
-      console.log(error);
-      return res.status(400).json({ type: 'error', response: error });
-    }
+    return res.status(200).json({ type: 'success' });
   }
 }
